Moved mapreduce imports to be from the sub-module, not individual files.
