+++ src/MEDS_transforms/extract/shard_events.py
@@ -15,7 +15,7 @@ from meds import subject_id_field
 from omegaconf import DictConfig, OmegaConf
 
 from MEDS_transforms.extract import CONFIG_YAML
-from MEDS_transforms.mapreduce.mapper import rwlock_wrap
+from MEDS_transforms.mapreduce import rwlock_wrap
 from MEDS_transforms.utils import get_shard_prefix, is_col_field, parse_col_field, write_lazyframe
 
 logger = logging.getLogger(__name__)
