+++ src/MEDS_transforms/filters/filter_measurements.py
@@ -7,7 +7,7 @@ import polars as pl
 from omegaconf import DictConfig
 
 from MEDS_transforms import PREPROCESS_CONFIG_YAML
-from MEDS_transforms.mapreduce.mapper import map_over
+from MEDS_transforms.mapreduce import map_over
 
 
 def filter_measurements_fntr(
