+++ src/MEDS_transforms/transforms/occlude_outliers.py
@@ -7,7 +7,7 @@ import polars as pl
 from omegaconf import DictConfig
 
 from MEDS_transforms import PREPROCESS_CONFIG_YAML
-from MEDS_transforms.mapreduce.mapper import map_over
+from MEDS_transforms.mapreduce import map_over
 
 
 def occlude_outliers_fntr(
